requirements.txt
setuptools>=65.5.1
wheel>=0.38.4
numpy>=1.26.0
flask==2.3.2
pandas==2.2.0
openpyxl==3.1.2
requests==2.31.0
langdetect==1.0.9
pyjwt==2.8.0
gunicorn==21.2.0
cryptography>=39.0.0
python-dotenv==1.0.0
chardet==5.2.0
openai>=2.3.0